+++ src/importer.js
@@ -15,7 +15,7 @@ import Solver from './solver';
  * @param {Solver} [solver]
  * @returns {Solver}
  */
-function importer_main(str, solver) {
+function importer_main(str, solver, _debug) {
   if (!solver) solver = new Solver();
 
   let pointer = 0;
@@ -120,11 +120,43 @@ function importer_main(str, solver) {
   }
 
   function parseIdentifier() {
+    if (read() === '\'') return parseQuotedIdentifier();
+    else return parseUnquotedIdentifier();
+  }
+
+  function parseQuotedIdentifier() {
+    is('\'');
+
+    let start = pointer;
+    while (!isEof() && read() !== '\'') skip();
+    if (isEof()) THROW('Quoted identifier must be closed');
+    if (start === pointer) THROW('Expected to parse identifier, found none');
+    skip(); // quote
+    return str.slice(start, pointer - 1); // return unquoted ident
+  }
+
+  function parseUnquotedIdentifier() {
     // anything terminated by whitespace
     let start = pointer;
-    while (!isEof() && !isWhite(read()) && read() !== '(' && read() !== ')' && read() !== ',') skip();
+    if (read() >= '0' && read() <= '9') THROW('Unquoted ident cant start with number');
+    while (!isEof() && isValidUnquotedIdentChar(read())) skip();
+    if (isEof()) THROW('Quoted identifier must be closed');
     if (start === pointer) THROW('Expected to parse identifier, found none');
     return str.slice(start, pointer);
+  }
+  function isValidUnquotedIdentChar(c) {
+    switch (c) {
+      case '(':
+      case ')':
+      case ',':
+      case '[':
+      case ']':
+      case '\'':
+      case '#':
+        return false;
+    }
+    if (isWhite(c)) return false;
+    return true;
   }
 
   function parseAlias() {
@@ -692,7 +724,8 @@ function importer_main(str, solver) {
         case 'val-strat':
           parseValStrat();
           break;
-        case 'set-markov':
+        case 'set-valdist':
+          skipWhitespaces();
           let target = parseIdentifier();
           let config = parseRestCustom();
           solver.setValueDistributionFor(target, JSON.parse(config));
@@ -770,6 +803,9 @@ function importer_main(str, solver) {
   }
 
   function THROW(msg) {
+    if (_debug) {
+      console.log(str.slice(0, pointer) + '##|PARSER_IS_HERE[' + msg + ']|##' + str.slice(pointer));
+    }
     msg += ', source at #|#: `' + str.slice(Math.max(0, pointer - 20), pointer) + '#|#' + str.slice(pointer, Math.min(str.length, pointer + 20)) + '`';
     throw new Error(msg);
   }